pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "payment-graph-forecasting"
version = "0.1.0"
description = "Forecasting the dynamics of the payment graph based on open payment data"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "pyarrow>=12.0.0",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
    "joblib>=1.3.0",
    "PyYAML>=6.0.0",
]

[project.optional-dependencies]
baselines = [
    "scikit-learn>=1.3.0",
    "catboost>=1.2.0",
    "statsmodels>=0.14.0",
    "prophet>=1.1.0",
    "matplotlib>=3.7.0",
]
dl = [
    "torch>=2.0.0",
    "ninja>=1.11.0",
]
pairmlp-cpu = [
    # All deps needed to run precompute.py on a CPU-only machine.
    # Install: pip install -e ".[pairmlp-cpu]"
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "pyarrow>=12.0.0",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
    "joblib>=1.3.0",
]
pairmlp-gpu = [
    # All deps needed to run run.py on a GPU machine (includes pairmlp-cpu deps).
    # Install: pip install -e ".[pairmlp-gpu]"
    # Then separately: pip install torch --index-url https://download.pytorch.org/whl/cu121
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "pyarrow>=12.0.0",
    "requests>=2.28.0",
    "tqdm>=4.65.0",
    "joblib>=1.3.0",
]
hpo = [
    "optuna>=3.0.0",
]
viz = [
    "matplotlib>=3.7.0",
    "networkx>=3.0",
]
dev = [
    "pytest>=7.0.0",
]
all = [
    "payment-graph-forecasting[baselines,dl,hpo,viz,dev]",
]

[tool.setuptools.packages.find]
include = ["payment_graph_forecasting*", "src*", "sg_baselines*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
